requirements.txt
openai==0.27.8
tiktoken==0.4.0
python-telegram-bot==20.5
